+++ src/serac/physics/utilities/functional/tests/CMakeLists.txt
@@ -61,7 +61,8 @@ endforeach()
 if(ENABLE_CUDA)
 
     set(functional_tests_cuda
-        tensor_unit_tests_cuda.cu)
+        tensor_unit_tests_cuda.cu
+	functional_comparisons_cuda.cu)	
         
     foreach(filename ${functional_tests_cuda})
         get_filename_component(test_name ${filename} NAME_WE)
